curve_name moved to mod tests
