+++ src/utils.rs
@@ -51,9 +51,13 @@ pub fn z_of_set<'a, F: FftField>(xs: impl IntoIterator<Item=&'a F>) -> DensePoly
 }
 
 #[cfg(test)]
-pub fn curve_name<E: PairingEngine>() -> &'static str {
-    // ark_ec::models::bw6::BW6<ark_bw6_761::curves::Parameters>
-    let full_name = std::any::type_name::<E>();
-    full_name.split_once("<").unwrap().1
-        .split_once(":").unwrap().0
+pub mod tests {
+    use ark_ec::PairingEngine;
+
+    pub fn curve_name<E: PairingEngine>() -> &'static str {
+        // ark_ec::models::bw6::BW6<ark_bw6_761::curves::Parameters>
+        let full_name = std::any::type_name::<E>();
+        full_name.split_once("<").unwrap().1
+            .split_once(":").unwrap().0
+    }
 }
